samples/mcp-dynamic-server/pyproject.toml
[project]
name = "mcp-dynamic-server"
version = "0.0.7"
description = "Dynamic tools MCP Server"
authors = [{ name = "John Doe" }]
dependencies = [
    "uipath-mcp==0.0.77",
]
requires-python = ">=3.10"
